backend/requirements.txt
# Campus Network Traffic Analyzer - Backend Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
scapy>=2.5.0
mysql-connector-python>=8.2.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pyyaml>=6.0.2
